+++ platform/jewel/ui/src/main/kotlin/org/jetbrains/jewel/ui/component/SegmentedControl.kt
@@ -27,6 +27,7 @@ import androidx.compose.ui.zIndex
 import org.jetbrains.jewel.foundation.GenerateDataFunctions
 import org.jetbrains.jewel.foundation.Stroke
 import org.jetbrains.jewel.foundation.modifier.border
+import org.jetbrains.jewel.foundation.modifier.thenIf
 import org.jetbrains.jewel.foundation.state.CommonStateBitMask
 import org.jetbrains.jewel.foundation.state.FocusableComponentState
 import org.jetbrains.jewel.foundation.theme.JewelTheme
@@ -35,7 +36,6 @@ import org.jetbrains.jewel.ui.component.styling.SegmentedControlStyle
 import org.jetbrains.jewel.ui.focusOutline
 import org.jetbrains.jewel.ui.theme.segmentedControlButtonStyle
 import org.jetbrains.jewel.ui.theme.segmentedControlStyle
-import org.jetbrains.jewel.ui.util.thenIf
 
 @Composable
 public fun SegmentedControl(
