+++ platform/jewel/ui/src/main/kotlin/org/jetbrains/jewel/ui/util/ModifierExtensions.kt
@@ -1,6 +1,7 @@
 package org.jetbrains.jewel.ui.util
 
 import androidx.compose.ui.Modifier
+import org.jetbrains.annotations.ApiStatus
 
 /**
  * Conditionally applies the [action] to the receiver [Modifier], if [precondition] is true. Returns the receiver as-is
@@ -8,7 +9,8 @@ import androidx.compose.ui.Modifier
  */
 @Deprecated(
     message = "This modifier has been moved to org.jetbrains.jewel.foundation.util. Please update your imports.",
-    ReplaceWith("thenIf(precondition, action)", "org.jetbrains.jewel.foundation.util"),
+    ReplaceWith("thenIf(precondition, action)", "org.jetbrains.jewel.foundation.modifier.thenIf"),
 )
+@ApiStatus.ScheduledForRemoval(inVersion = "2025.2")
 public inline fun Modifier.thenIf(precondition: Boolean, action: Modifier.() -> Modifier): Modifier =
     if (precondition) action() else this
